Implement short screen mode

+++ include/sdl_utils.h
@@ -10,8 +10,7 @@ constexpr int pow2(const int number) {
 
 std::string find_sans_serif_font_path();
 
-void convert_bw1bit_to_rgb888(const std::span<uint8_t> &bw1bit_buf,
-                              std::vector<uint32_t> &rgb888_buf);
+void convert_bw1bit_to_rgb888(const std::span<uint16_t> &bw1bit_buf, std::vector<uint32_t> &rgb888_buf);
 
 void convert_bgr565_to_rgb888(const std::span<uint16_t> &bgr565_buf, std::vector<uint32_t> &rgb888_buf);
